requirements.txt
graphql_core==2.3.2
Django==3.0.5
graphene_django==2.15.0
PyJWT==1.7.0
django_graphql_jwt==0.3.1
graphene==2.1.8

